+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [**#71**](https://github.com/psake/PowerShellBuild/pull/71) Compiled modules
   are now explicitly created as UTF-8 files.
-- [**#67**](https://github.com/psake/PowerShellBuild/pull/67) Overwrite existing
-  markdown files and use the comment based help as the source of truth. This can
-  be disabled if `$PSBPreference.Docs.Overwrite` is set to `$false`.
+- [**#67**](https://github.com/psake/PowerShellBuild/pull/67) You can now
+  overwrite existing markdown files using `$PSBPreference.Docs.Overwrite` and
+  setting it to `$true`.
 
 ## [0.6.2] 2024-10-06